[Unity] Value System Script (C#)
A C# script used in a small Unity game I created about completing tasks around a house in quick succession. This was the script that handled the adjustments to task timers and delays, alongside increasing score earnings over time, making the game get faster over time as a form of difficulty and reward.