[Unity] TV Task System Script (C#)
A C# script used in a small Unity game I created about completing tasks around a house in quick succession. This was one of four scripts that handled the functionality of each of the tasks I implemented into the game. This one was by far, the longest script of the four.